backend/requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
fastapi==0.111.0
gunicorn==22.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
packaging==24.0
pony==0.7.17
pycparser==2.22
PyMySQL==1.1.0
typing_extensions==4.11.0
uvicorn==0.29.0
Werkzeug==3.0.3
python-jose==3.3.0
passlib==1.7.4
